hide the sharing for URLs containing 'onsite'
It's a little hacky but I didn't want to pollute the editor with an
option for this yet. So now whenever a story URL contains 'onsite' it
will hide the sharing buttons.

If this becomes a regularly used feature it would be better to
officially add this as an option for the entry screen.