limit the amount of navigation pages
we've seen stories to go beyond this. While in general folks should use
the 'show in navigation' checkbox, it's a good idea to limit this
server-side as well, in case anyone forgets.

If important pages are chopped of the end, bringing them down using the
checkboxes should then bring it under 25.